requirements.txt
aiohttp
async_generator
bs4
cowpy
cffi
coffeehouse
cryptg
dnspython
emoji
gitpython
glitch_this
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google_images_download
google_trans_new
gTTS
gTTS-token
hachoir
heroku3
html_telegraph_poster
httplib2
humanize
jikanpy
lxml
lyricsgenius
oauth2client
pendulum
Pillow
psutil
psycopg2
psycopg2-binary
pybase64
pySmartDL
python-barcode
python-dotenv
youtube-dl
pytz
redis
requests
search-engine-parser
speedtest-cli
sqlalchemy
TgCrypto
telethon
telethon-session-sqlalchemy
telegraph
urbandict
PyGithub
python-dateutil
vcsi
